# api-spec/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    CommentAuthorResponse:
      type: object
      examples:
        - id: user-1
          username: PlayerOne
      description: Author information for a comment
      properties:
        id:
          type: string
          examples:
            - user-1
          description: Unique author identifier
        username:
          type: string
          examples:
            - PlayerOne
          description: Author username
    CommentResponse:
      type: object
      examples:
        - id: "117457749108987475"
          author:
            id: user-1
            username: PlayerOne
          content: Great feedback!
          isDeveloperResponse: false
          upvotes: 5
          createdAt: 2024-01-15T10:30:00Z
          updatedAt: 2024-01-15T10:30:00Z
      description: Feedback comment details response
      properties:
        id:
          type: string
          examples:
            - "117457749108987475"
          description: Unique comment identifier
        author:
          $ref: "#/components/schemas/CommentAuthorResponse"
          type: object
          description: Author information
        content:
          type: string
          examples:
            - Great feedback!
          description: Comment content text
        isDeveloperResponse:
          type: boolean
          examples:
            - false
          description: Whether this is an official developer response
        upvotes:
          type: integer
          format: int32
          examples:
            - 5
          description: Number of upvotes on this comment
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
          type: string
          examples:
            - 2024-01-15T10:30:00Z
          description: Timestamp when comment was created
        updatedAt:
          $ref: "#/components/schemas/OffsetDateTime"
          type: string
          examples:
            - 2024-01-15T10:30:00Z
          description: Timestamp when comment was last updated
    CreateCommentRequest:
      type: object
      required:
        - text
      examples:
        - text:
            Great feedback! This is a valuable feature request that many users have
            been asking for.
      description: Request to create a new comment on feedback
      properties:
        text:
          type: string
          minLength: 1
          examples:
            - Great feedback! This is a valuable feature request that many users have
              been asking for.
          maxLength: 5000
          pattern: \S
          description: Comment text content
    CreateFeedbackCommand:
      type: object
      required:
        - title
        - description
      examples:
        - title: Dark mode support needed
          description:
            Users are requesting a dark mode option for better usability
            in low-light environments. This would reduce eye strain and improve overall
            user experience.
          categoryId: 117457749108987394
          authorId: 117457749108987388
          sentiment: POSITIVE
          tags: "ui,accessibility,enhancement"
      description: Request to create a new feedback item
      properties:
        title:
          type: string
          minLength: 3
          examples:
            - Dark mode support needed
          maxLength: 255
          pattern: \S
          description: Feedback title
        description:
          type: string
          minLength: 10
          examples:
            - Users are requesting a dark mode option for better usability in low-light
              environments. This would reduce eye strain and improve overall user experience.
          maxLength: 5000
          pattern: \S
          description: Detailed feedback description
        categoryId:
          type: integer
          format: int64
          examples:
            - 117457749108987394
          description: Category ID for feedback classification
        authorId:
          type: integer
          format: int64
          examples:
            - 117457749108987388
          description: Author/user ID who submitted the feedback
        sentiment:
          type: string
          enum:
            - POSITIVE
            - NEGATIVE
            - NEUTRAL
          examples:
            - POSITIVE
          description: Sentiment of the feedback
        tags:
          type: string
          examples:
            - "ui,accessibility,enhancement"
          description: Comma-separated tags for categorization
    FeedbackResponse:
      type: object
      examples:
        - id: "117457749108987399"
          title: Dark mode support needed
          description:
            Users are requesting a dark mode option for better usability
            in low-light environments.
          sentiment: POSITIVE
          upvotes: 45
          comments: 8
          status: ACKNOWLEDGED
          categoryName: Feature Request
          authorName: Bob Smith
          createdAt: 2025-11-20T11:30:00+00:00
          archived: false
      description: Feedback item details response
      properties:
        id:
          type: string
          examples:
            - "117457749108987399"
          description: Unique feedback identifier
        title:
          type: string
          examples:
            - Dark mode support needed
          description: Feedback title
        description:
          type: string
          examples:
            - Users are requesting a dark mode option for better usability in low-light
              environments.
          description: Detailed feedback description
        sentiment:
          type: string
          enum:
            - POSITIVE
            - NEGATIVE
            - NEUTRAL
          examples:
            - POSITIVE
          description: Sentiment of the feedback
        upvotes:
          type: integer
          format: int32
          examples:
            - 45
          description: Number of upvotes
        comments:
          type: integer
          format: int32
          examples:
            - 8
          description: Number of comments
        status:
          $ref: "#/components/schemas/FeedbackStatus"
          type: string
          enum:
            - PENDING
            - ACKNOWLEDGED
            - IN_PROGRESS
            - COMPLETED
          examples:
            - ACKNOWLEDGED
          description: Current status of the feedback
        categoryName:
          type: string
          examples:
            - Feature Request
          description: Category name for this feedback
        authorName:
          type: string
          examples:
            - Bob Smith
          description: Name of the feedback author
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
          type: string
          examples:
            - 2025-11-20T11:30:00+00:00
          description: Timestamp when feedback was created
        archived:
          type: boolean
          examples:
            - false
          description: Whether the feedback is archived
    FeedbackStatus:
      type: string
      enum:
        - PENDING
        - ACKNOWLEDGED
        - IN_PROGRESS
        - COMPLETED
    OffsetDateTime:
      type: string
      format: date-time
      examples:
        - 2022-03-10T12:15:50-04:00
    PaginatedFeedbackResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackResponse"
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
        total_items:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
    UpdateFeedbackCommand:
      type: object
      required:
        - title
        - description
        - status
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
          pattern: \S
        description:
          type: string
          minLength: 10
          maxLength: 5000
          pattern: \S
        status:
          $ref: "#/components/schemas/FeedbackStatus"
        categoryId:
          type: integer
          format: int64
        authorId:
          type: integer
          format: int64
        sentiment:
          type: string
        tags:
          type: string
    VoteRequest:
      type: object
      required:
        - direction
      properties:
        direction:
          type: string
          enum:
            - up
            - down
            - none
          examples:
            - up
          description: Vote direction (up for upvote, down for downvote, none to remove vote)
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: http://localhost:62505/realms/quarkus/.well-known/openid-configuration
...
paths:
...
  /api/v1/feedback:
    get:
      summary: List all feedback items with pagination and sorting
      description:
        Retrieve feedback submissions with support for pagination and sorting
        by creation date
      tags:
        - Feedback
      parameters:
        - description: "Page number (1-indexed, default 1)"
          example: 1
          name: page
          in: query
          schema:
            type: integer
            format: int32
        - description: "Page size (default 10, max 20)"
          example: 10
          name: pageSize
          in: query
          schema:
            type: integer
            format: int32
        - description: "Sort order: 'newest' (default) or 'oldest'"
          example: newest
          name: sortBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of feedback items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFeedbackResponse"
    post:
      summary: Submit new feedback
      description: Create a new feedback item
      tags:
        - Feedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackCommand"
        required: true
      responses:
        "201":
          description: Feedback created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: Invalid request - validation errors
  /api/v1/feedback/{id}:
    patch:
      summary: Update feedback
      description:
        Update an existing feedback item. Only the feedback author or an
        admin can update. Requires authentication via JWT token.
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedbackCommand"
        required: true
      responses:
        "200":
          description: Feedback updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: Invalid request - validation errors
        "403":
          description:
            Forbidden - either not authenticated or user lacks permission
            to update this feedback
        "404":
          description: Feedback not found
    get:
      summary: Get feedback by ID
      description: Retrieve a specific feedback item by its ID
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Feedback details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "404":
          description: Feedback not found
    delete:
      summary: Delete feedback
      description: Delete a feedback item by ID
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Feedback deleted successfully
        "404":
          description: Feedback not found
  /api/v1/feedback/{id}/archive:
    post:
      summary: Archive feedback
      description: "Mark a feedback item as archived, making it inactive without deletion"
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Feedback archived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "404":
          description: Feedback not found
  /api/v1/feedback/{id}/comments:
    get:
      summary: Get comments for feedback
      description: Retrieve all comments for a specific feedback item
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      responses:
        "200":
          description: List of comments retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "404":
          description: Feedback not found
    put:
      summary: Add comment to feedback
      description: Add a new comment to an existing feedback item
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
        required: true
      responses:
        "201":
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "400":
          description: Invalid request - validation errors
        "404":
          description: Feedback not found
  /api/v1/feedback/{id}/reopen:
    post:
      summary: Reopen feedback
      description: "Reopen an archived feedback item, making it active again"
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Feedback reopened successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "404":
          description: Feedback not found
  /api/v1/feedback/{id}/upvote:
    post:
      summary: Vote on feedback
      description: Upvote, downvote, or remove vote from feedback
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
        required: true
      responses:
        "200":
          description: Vote recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: Invalid vote direction
        "404":
          description: Feedback not found
  /api/v1/feedback/{id}/comments/{commentId}/upvote:
    post:
      summary: Vote on comment
      description: Upvote, downvote, or remove vote from comment
      tags:
        - Feedback
      parameters:
        - description: Feedback ID
          required: true
          name: id
          in: path
          schema:
            type: string
        - description: Comment ID
          required: true
          name: commentId
          in: path
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
        required: true
      responses:
        "200":
          description: Vote recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "400":
          description: Invalid vote direction
        "404":
          description: Comment or feedback not found
...
info:
  title: agora-feedback API
  version: 1.0-SNAPSHOT